+++ lib/domain/record/components/notification_bar/notification_bar.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:pilll/analytics.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/database/pilll_ads.dart';
+import 'package:pilll/provider/pilll_ads.dart';
 import 'package:pilll/domain/premium_introduction/premium_introduction_sheet.dart';
 import 'package:pilll/domain/premium_introduction/util/discount_deadline.dart';
 import 'package:pilll/domain/record/components/notification_bar/components/discount_price_deadline.dart';
