+++ test/domain/record/notification_bar/notification_bar_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:pilll/analytics.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/database/pilll_ads.dart';
+import 'package:pilll/provider/pilll_ads.dart';
 import 'package:pilll/domain/premium_introduction/util/discount_deadline.dart';
 import 'package:pilll/domain/record/components/notification_bar/components/discount_price_deadline.dart';
 import 'package:pilll/domain/record/components/notification_bar/components/ended_pill_sheet.dart';
